задание 3-1 правки
