+++ templates/main.php
@@ -28,7 +28,7 @@
           <div class="lot__state">
             <div class="lot__rate">
               <span class="lot__amount">Стартовая цена</span>
-              <span class="lot__cost"><?= formatAmount(esc($value['price'])); ?></span>
+              <span class="lot__cost"><?= esc(formatAmount($value['price'])); ?></span>
             </div>
             <div class="lot__timer timer">
               12:23
